requirements.txt
black
flake8
isort
mkdocs
pip
python-dotenv
openai
python-pptx
openai-whisper
-e .
